Confirmed hardware test with Raspberry Pi and BBB Connection Board Rev B
