+++ src/main/java/org/karnes/homebrew/hardware/MCP23017.java
@@ -146,43 +146,6 @@ public class MCP23017 {
 
 }
 
-public enum MCP23017Pin {
-
-    GPIOA_0(MCP23017Port.A, (byte) 0),
-    GPIOA_1(MCP23017Port.A, (byte) 1),
-    GPIOA_2(MCP23017Port.A, (byte) 2),
-    GPIOA_3(MCP23017Port.A, (byte) 3),
-    GPIOA_4(MCP23017Port.A, (byte) 4),
-    GPIOA_5(MCP23017Port.A, (byte) 5),
-    GPIOA_6(MCP23017Port.A, (byte) 6),
-    GPIOA_7(MCP23017Port.A, (byte) 7),
-
-    GPIOB_0(MCP23017Port.B, (byte) 0),
-    GPIOB_1(MCP23017Port.B, (byte) 1),
-    GPIOB_2(MCP23017Port.B, (byte) 2),
-    GPIOB_3(MCP23017Port.B, (byte) 3),
-    GPIOB_4(MCP23017Port.B, (byte) 4),
-    GPIOB_5(MCP23017Port.B, (byte) 5),
-    GPIOB_6(MCP23017Port.B, (byte) 6),
-    GPIOB_7(MCP23017Port.B, (byte) 7);
-
-    private final MCP23017Port port;
-    private final byte position;
-
-    MCP23017Pin(MCP23017Port port, byte position) {
-        this.port = port;
-        this.position = position;
-    }
-
-    public byte getPosition() {
-        return position;
-    }
-
-    public MCP23017Port getPort() {
-        return port;
-    }
-}
-
 enum MCP23017Port {
     A((short) 0), B((short) 1);
 
@@ -197,24 +160,3 @@ enum MCP23017Port {
     }
 }
 
-public enum MCP23017Address {
-    ADDR0(0x20),
-    ADDR1(0x21),
-    ADDR2(0x22),
-    ADDR3(0x23),
-    ADDR4(0x24),
-    ADDR5(0x25),
-    ADDR6(0x26),
-    ADDR7(0x27);
-
-    //Because the underlying library wants shorts
-    private final short hardwareValue;
-
-    MCP23017Address(int hardwareValue) {
-        this.hardwareValue = (short) hardwareValue;
-    }
-
-    public short getHardwareValue() {
-        return hardwareValue;
-    }
-}
